control/requirements.txt
IPython[cv]
ipykernel[cv]
camelot-py[cv]
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
ghostscript==0.7
Levenshtein==0.20.9
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
pandas==1.5.3
pdfminer.six==20221105
pycparser==2.21
pycryptodome==3.17
PyPDF2==2.10.3
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.7.1
rapidfuzz==2.13.7
six==1.16.0